added parent window handling to rhino
also cleaned up and reorganized solution folder
